// src/containers/member-register/type.ts
export type IFormInput = {
	companyName: string
	companyNameFurigana: string
	companyCode: string
	representerFirstName: string
	representerLastName: string
	representerFirstNameFurigana: string
	representerLastNameFurigana: string
	postCode: string
	province: string
	city: string
	address: string
	firstPhone: string
	middlePhone: string
	lastPhone: string
	monthFinance: number
	personInChargeFirstName: string
	personInChargeLastName: string
	personInChargeFirstNameFurigana: string
	personInChargeLastNameFurigana: string
	hasAdvisoryTax: boolean
	isGoodReporting: boolean
	isCorrespondCashHandling: boolean
	email: string
	reEnterEmail: string
	password: string
	reEnterPassword: string
}